Make sdk2013-mp build on x64 platform (#2)
* Make it build on x64 platform

* Add linux64 libraries

* Use ThreadId_t

* Update libtier0.so to latest version

It was one commit behind.
There is a newer one from 8 months ago:
https://github.com/ValveSoftware/source-sdk-2013/blob/master/src/lib/public/linux64/libtier0.so

---------

+++ sdk2013-mp/src/tier1/strtools.cpp
@@ -2001,7 +2001,7 @@ bool V_ExtractFilePath (const char *path, char *dest, int destSize )
 		src--;
 	}
 
-	int copysize = min( src - path, destSize - 1 );
+	int copysize = min( (int)(src - path), destSize - 1 );
 	memcpy( dest, path, copysize );
 	dest[copysize] = 0;
